Refactoring and code separations.
